+++ tests/testthat/test-config.R
@@ -52,3 +52,21 @@ test_that("functions are returned as-is for '*' types", {
   })
 
 })
+
+test_that("we can query options without warnings", {
+
+  local({
+    renv_scope_envvars(RENV_CONFIG_EXTERNAL_LIBRARIES = "/tmp")
+    expect_identical(config$external.libraries(), "/tmp")
+  })
+
+})
+
+test_that("invalid configuration options trigger a warning", {
+
+  local({
+    renv_scope_options(renv.config.connect.timeout = "oops")
+    expect_warning(config$connect.timeout())
+  })
+
+})